update tests to increase code coverage
By this change it is possible to note that parts of the code may
not be tested depending on the date the tests are running (e.g. a
test that needs to test the path of a date one year ago but next
month, but with the test being executed in the month of December.
In this case, instead of testing a date in the previous year, the
test would be actually going through the path of the current year,
because when summing one month to December we go to January of next
Year.
This is a code debt that will be addressed later.