bug fix and removal of code duplication
The bug fixed is for the case where the calculated  birth date is
in the month of January. In the old implementation, by using
!birthDate.getMonth() it would return true if the returned month is
equal to zero '0', and since months in JavaScript start in zero and
end in eleven this was a valid case, but in the case of a calcula-
ted month being in January, it is incorrect going through this
path.